+++ BookDinner/Class/Custom/ToolBar/BDToolBar.h
@@ -11,8 +11,8 @@
 
 @interface BDToolBar : HNYView
 @property (nonatomic,strong) NSMutableArray *subMenuAry;
-@property (nonatomic) int selectIndex;
-@property (nonatomic) int defaultSelectedIndex;
+@property (nonatomic) NSInteger selectIndex;
+@property (nonatomic) NSInteger defaultSelectedIndex;
 
 
 @end
