+++ BookDinner/Class/Person/BDPersonDetailViewController.m
@@ -300,7 +300,7 @@
 - (void)hNYPopoverView:(HNYPopoverView *)popover didSelectStringAryAtIndex:(NSInteger)index{
     HNYDetailItemModel *item = [self.tableViewController getItemWithKey:KUSER_SEX];
     item.textValue = [self.sexAry objectAtIndex:index];
-    item.value = [NSString stringWithFormat:@"%d",index];
+    item.value = [NSString stringWithFormat:@"%ld",(long)index];
     self.sexTextField.text = item.textValue;
     [popover dismissPopoverAnimated:YES];
 }
