+++ BookDinner/Common/TableList/HNYTableView.m
@@ -27,7 +27,7 @@
     return 1;
 }
 
-- (int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section{
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section{
     return self.list.count;
 }
